feat(react): add script to copy normalize.css, update React instructions
Update scripts in package.json to copy both core and normalize CSS files separately to React
package, update README and Getting Started instructions to include normalize.css guidance, add
normalize.css import to preview.js Storybook file.

BREAKING CHANGE: Normalised styles are no longer exported as part of the core.css file. In projects
which require the same normalised styles as are used within the components to be applied to the rest
of the project or slotted elements within the components, a separate import for the normalize.css
file will now be needed.