feat(web-components): export normalize.css separately from core.css, updated instructions
Update Stencil config to import normalize.css alongside core.css instead of including it within
core.css, update instructions about using styles to inform about the normalize.css import, rename
'normalise' to American spelling (to match spelling of Normalize.css library and how people are
usually used to seeing it spelt in code), import normalize.css into Storybook (preview.js) to
apply normalised styles to slotted elements.

BREAKING CHANGE: Normalised styles are no longer exported as part of the core.css file. In projects
which require the same normalised styles as are used within the components to be applied to the rest
of the project or slotted elements within the components, a separate import for the normalize.css
file will now be needed.