feat: fix vertical spacing in ic-page-header
Removed row gap from header container, as the only 2 elements in this were the heading and the
adornment. The heading has natural padding due to the line height of the h2 text. Added margin
bottom to keep space between adornment and subheading. This fix is for small or mobile sized
screens.