feat(canary-web-components): added checkboxes to perform row selection in ic-data-table
refactored `highlight-selected-row` to default to false. When true, it will display checkboxes for
each row, that highlight the row when checked. the row selection event will also emit the array of
all selected rows, since multiple rows can be selected at once. Added story for this, removing old
stories for previous functionality

BREAKING CHANGE: The highlight-selected-row prop now defaults to false, with the functionality of
clicking a row's cell to select it now disabled. The only way to select a row at the moment is to
click the checkbox. The event detail for row selection has been updated to an object with two
values, instead of one value