fix(web-components): updated buttons prop name in ic-dialog to hide-default-controls
updated prop name to be more descriptive of purpose, with default value now being `false`. Have
adjusted logic to not hide controls if the prop is true but there is slotted content present.
Updated unit tests and stories. Minor refactor of code.

BREAKING CHANGE: buttons prop is now hide-default-controls, with a default value of false.