fix(web-components): updated hover/pressed colours for tertiary/nav buttons
updated the hover/pressed background to correct values. Updated nav button logic where the theme is
updated and applied to the host element, otherwise the correct colours don't get passed down.
Updated its variant to icon-tertiary which feels more appropriate, but also added the variant-icon
class to the button when any of the icon variants are used so it receives correct styling