feat(root): included angular output targets to produce native wrappers
Set up new output targets in the stencil config to output native angular wrapper components in a
similar vein to the react components. Versions have been included that will allow the integrating
developers to import the components as standalone items one-by-one at need or as a whole module.
This also required the addition/alteration of the custom-elements-dist output to create an internal
package for the web-components that the angular output plugin could use.

re #3933